requirements.txt
flask
flask-cors
numpy
pandas
plotly
gunicorn
python-dotenv
